+++ src/helpers.ts
@@ -4,26 +4,56 @@ import {LoggerConfig, loggerLevels} from './config-type';
 import {MiddlewareDefinition} from './middleware';
 
 export type FlatObject = Record<string, any>;
-export type ObjectSchema = z.AnyZodObject;
+type EmptyFlatObject = {[K in never]: never};
 
-export type Merge<A extends ObjectSchema, B extends ObjectSchema | any> = z.ZodObject<
-  // eslint-disable-next-line @typescript-eslint/ban-types
-  A['shape'] & (B extends ObjectSchema ? B['shape'] : {}),
-  A['_unknownKeys'],
-  A['_catchall']
+type ObjectSchema = z.AnyZodObject;
+type Extractable = 'shape' | '_unknownKeys' | '_catchall';
+type UnionSchema = z.ZodUnion<[ObjectSchema, ...ObjectSchema[]]>;
+type IntersectionSchema = z.ZodIntersection<ObjectSchema, ObjectSchema>;
+
+export type IOSchema = ObjectSchema | UnionSchema | IntersectionSchema;
+
+type ArrayElement<T extends readonly unknown[]> = T extends readonly (infer K)[] ? K : never;
+
+type UnionResult<T extends ObjectSchema[], F extends Extractable> = Partial<ArrayElement<T>[F]>;
+type IntersectionResult<T extends IntersectionSchema, F extends Extractable> =
+  T['_def']['left'][F] & T['_def']['right'][F];
+
+type IOExtract<T extends IOSchema | any, F extends Extractable> =
+  T extends ObjectSchema ? T[F] :
+  T extends UnionSchema ? UnionResult<T['options'], F> :
+  T extends IntersectionSchema ? IntersectionResult<T, F> : EmptyFlatObject;
+
+export type Merge<A extends IOSchema, B extends IOSchema | any> = z.ZodObject<
+  IOExtract<A, 'shape'> & IOExtract<B, 'shape'>,
+  IOExtract<A, '_unknownKeys'>,
+  IOExtract<A, '_catchall'>
 >;
 
-export function combineEndpointAndMiddlewareInputSchemas<IN extends ObjectSchema, mIN>(
+export function extractObjectSchema(subject: IOSchema): ObjectSchema {
+  if (subject instanceof z.ZodUnion) {
+    return subject.options.reduce((acc, option) =>
+      acc.partial().merge(option.partial()));
+  }
+  if (subject instanceof z.ZodIntersection) {
+    return subject._def.left.merge(subject._def.right);
+  }
+  return subject;
+}
+
+export function combineEndpointAndMiddlewareInputSchemas<IN extends IOSchema, mIN>(
   input: IN,
-  middlewares: MiddlewareDefinition<any, any, any>[]
+  middlewares: MiddlewareDefinition<IOSchema, any, any>[]
 ): Merge<IN, mIN> {
   if (middlewares.length === 0) {
-    return input as any as Merge<IN, mIN>;
+    return extractObjectSchema(input) as Merge<IN, mIN>;
   }
-  return middlewares
+  const mSchema = middlewares
     .map((middleware) => middleware.input)
-    .reduce((carry: ObjectSchema, schema) => carry.merge(schema))
-    .merge(input) as Merge<IN, mIN>;
+    .reduce((carry, schema) =>
+      extractObjectSchema(carry).merge(extractObjectSchema(schema))
+    );
+  return extractObjectSchema(mSchema).merge(extractObjectSchema(input)) as Merge<IN, mIN>;
 }
 
 export function getInitialInput(request: Request): any {