+++ src/open-api.ts
@@ -5,6 +5,7 @@ import {
 } from 'openapi3-ts';
 import {ReferenceObject} from 'openapi3-ts/src/model/OpenApi';
 import {z} from 'zod';
+import {extractObjectSchema} from './helpers';
 import {Routing, routingCycle} from './routing';
 import {lookup} from 'mime';
 
@@ -188,14 +189,15 @@ export class OpenAPI {
       };
       if (method === 'get') {
         operation.parameters = [];
-        Object.keys(endpoint.getInputSchema().shape).forEach((name) => {
+        const subject = extractObjectSchema(endpoint.getInputSchema()).shape;
+        Object.keys(subject).forEach((name) => {
           const parameterRef = this.createRef('parameter', 'parameters');
           this.builder.addParameter(parameterRef.name, {
             name,
             in: 'query',
-            required: !endpoint.getInputSchema().shape[name].isOptional(),
+            required: !subject[name].isOptional(),
             schema: {
-              ...describeSchema(endpoint.getInputSchema().shape[name], false),
+              ...describeSchema(subject[name], false),
               description: `${fullPath} ${method.toUpperCase()} parameter`
             },
           });